pr feedback: better interface contract to avoid deprecated mono side implementation
